+++ fish/fish.h
@@ -69,7 +69,7 @@ extern int issue_command (const char *cmd, char *argv[], const char *pipe, int r
 extern void list_builtin_commands (void);
 extern int display_builtin_command (const char *cmd);
 extern void free_strings (char **argv);
-extern int count_strings (char *const *argv);
+extern size_t count_strings (char *const *argv);
 extern void print_strings (char *const *argv);
 extern void print_table (char *const *argv);
 extern int is_true (const char *str);
